Refactor React Components to stateless style
